+++ 1/app/components/src/firebaseConfig.js
